Converted to functional component structure, splitting into Parent-Children relationship to make code readable and move towards SOLID principles
