+++ components/Nav.tsx
@@ -2,18 +2,18 @@ import Link from 'next/link'
 import React from 'react'
 import Image from 'next/image'
 
-export default function Nav () {
+const Nav = () => {
   return (
     <nav className="navbar" role="navigation" aria-label="main navigation">
       <div className="navbar-brand">
-        <Link href="/" passHref>
+        <a href="/" >
           <Image
-            width="112"
-            height="40"
+            width="150"
+            height="90"
             alt="Website Logo"
             src="https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fbestbrokerindia.com%2Fblog%2Fwp-content%2Fuploads%2F2019%2F05%2FTechnical-Analysis-Basics.jpg&f=1&nofb=1"
           />
-        </Link>
+        </a>
 
         <a
           role="button"
@@ -30,14 +30,14 @@ export default function Nav () {
       <div id="navbarBasicExample" className="navbar-menu">
         <div className="navbar-start">
           <div className="navbar-item">
-            <Link passHref href="/">
-              <a>Home</a>
-            </Link>
+            <a href="/">
+              Home
+            </a>
           </div>
           <div className="navbar-item">
-            <Link passHref href="/chart">
-            <a>Charts</a>
-            </Link>
+            <a href="/chart">
+            Charts
+            </a>
           </div>
         </div>
 
@@ -55,3 +55,6 @@ export default function Nav () {
     </nav>
   )
 }
+
+
+export default Nav
